Adjusting bean validation

+++ example/src/main/java/com/github/braisdom/objsql/example/QueryExample.java
@@ -42,8 +42,8 @@ public class QueryExample {
             orders.add(new Order()
                     .setNo("20200000" + i)
                     .setMemberId(i)
-                    .setAmount(RandomUtils.nextFloat(10.0f, 30.0f))
-                    .setQuantity(RandomUtils.nextFloat(100.0f, 300.0f))
+                    .setAmount(RandomUtils.nextDouble(10.0, 30.0))
+                    .setQuantity(RandomUtils.nextDouble(100.0f, 300.0f))
                     .setSalesAt(Timestamp.valueOf("2020-05-01 09:30:00")));
         }
